+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1] - TODO: Add release date.
+# 0.1.0
 
-* TODO: Describe initial release.
+* Initial development release
+* Added *GMBAPI* as an entry point for connection with Google through Google Sign In
+* Added support for fetching locations from [GMB Locations API](https://developers.google.com/my-business/reference/rest/v4/accounts.locations)